Updates for better halo finding in parallel on cluster
A phony target is added for the halo finding, so that can be
done separately. ROCKSTAR needs multiple cores, so it should
be run through a PBS script on a cluster. But the rest is serial
and doesn't benefit from the extra cores, so we don't want to
waste time with those sections.