updates to the read_tree functionality
This is still not totally funcitonal. There are now checks to make
sure the merger is with the main halo. I also wrote the code to
trace backwards in time to find the maximum mass before accretion.
This is not tested, as the code to create the output files is
not complete yet.